# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.CreateAdminRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: Admin@123
        type: string
      phone:
        example: "13800138000"
        type: string
      username:
        example: admin123
        type: string
    required:
    - email
    - password
    - phone
    - username
    type: object
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
  controllers.UpdateAdminRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: NewPassword@123
        type: string
      phone:
        example: "13800138000"
        type: string
    type: object
...
host: localhost:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
  /admins:
    get:
      consumes:
      - application/json
      description: Get a list of all admin users with pagination
      parameters:
      - description: Page number, default is 1
        in: query
        name: page
        type: integer
      - description: Items per page, default is 10
        in: query
        name: page_size
        type: integer
      - description: Search keyword for username, phone, etc.
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Admin List
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new admin user
      parameters:
      - description: Admin information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Admin
      tags:
      - Admin
  /admins/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an admin user with the specified ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Admin
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get details of a specific admin by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Admin By ID
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update an existing admin user
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated admin information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Admin
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"